+++ mitra-models/migrations/schema.sql
@@ -162,6 +162,7 @@ CREATE TABLE media_attachment (
     file_name VARCHAR(200) NOT NULL,
     file_size INTEGER,
     media_type VARCHAR(50),
+    description VARCHAR(1500),
     ipfs_cid VARCHAR(200),
     post_id UUID REFERENCES post (id) ON DELETE CASCADE,
     created_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT now()