+++ src/mastodon_api/media/types.rs
@@ -12,6 +12,7 @@ pub struct AttachmentCreateData {
     // base64-encoded file (not comtaible with Mastodon)
     pub file: String,
     pub media_type: String,
+    pub description: Option<String>,
 }
 
 /// https://docs.joinmastodon.org/entities/attachment/
@@ -24,6 +25,7 @@ pub struct Attachment {
 
     pub url: String,
     pub preview_url: String,
+    description: Option<String>,
 }
 
 impl Attachment {
@@ -45,6 +47,7 @@ impl Attachment {
             attachment_type: attachment_type_mastodon.to_string(),
             url: attachment_url.clone(),
             preview_url: attachment_url,
+            description: db_attachment.description,
         }
     }
 }